requirements.txt
langchain
langchain-community
langchain-ollama
chromadb
chainlit
pymupdf
python-dotenv
fastapi
uvicorn
streamlit
faiss-cpu
rank_bm25
python-docx
docx2txt
openpyxl
